refactor: Change int and float to decimal

+++ Banking-System/Core/Models/Account.cs
@@ -10,8 +10,8 @@ namespace Banking_System.Core.Models
         public required string AccountNumber { get; set; }
         [Required]
         public required string AccountType { get; set; }
-        public int Balance { get; set; }
-        public int OverDraftLimit { get; set; }
+        public decimal Balance { get; set; }
+        public decimal OverDraftLimit { get; set; }
         public float IntrestRate { get; set; }
         public decimal AccumulatedInterest { get; set; }
         public DateTime LastInterestCalculationDate { get; set; }
